:hammer: Extract packages
