+++ packages/index.js
@@ -1,0 +1,60 @@
+// base
+import ULink from './u-link'
+import UButton from './u-button'
+import UIcon from './u-icon'
+import ULayout from './u-layout'
+
+// form
+import UInput from './u-input'
+import USelect from './u-select'
+import USwitch from './u-switch'
+import UCheckbox from './u-checkbox/u-checkbox'
+import UCheckboxs from './u-checkbox/u-checkboxs'
+import URadio from './u-radio/u-radio'
+import URadios from './u-radio/u-radios'
+import UForm from './u-form/u-form'
+import UFormItem from './u-form/u-form-item'
+
+// usage
+import UTableColumn from './u-table/u-table-column'
+import UTable from './u-table/u-table'
+import UPagination from './u-pagination'
+import UTab from './u-tab/u-tab'
+import UTabs from './u-tab/u-tabs'
+import UCrumb from './u-crumb/u-crumb.vue'
+import UCrumbItem from './u-crumb/u-crumb-item.vue'
+
+// popup
+import UModal from './u-modal'
+import UToast from './u-toast'
+import UPopper from './u-popper'
+import UTooltip from './u-tooltip'
+
+import './assets/css/reset.scss'
+
+export {
+    ULink,
+    UButton,
+    UInput,
+    USelect,
+    UCheckbox,
+    UCheckboxs,
+    USwitch,
+    ULayout,
+    UTableColumn,
+    UTable,
+    UPagination,
+    UTab,
+    UTabs,
+    UModal,
+    UToast,
+    UPopper,
+    UTooltip,
+    URadio,
+    URadios,
+    UIcon,
+    UForm,
+    UFormItem,
+    UCrumb,
+    UCrumbItem
+}
