+++ src/components/d-component-item.vue
