--- src/entry.js
@@ -1,60 +1,0 @@
-// base
-import ULink from './packages/u-link'
-import UButton from './packages/u-button'
-import UIcon from './packages/u-icon'
-import ULayout from './packages/u-layout'
-
-// form
-import UInput from './packages/u-input'
-import USelect from './packages/u-select'
-import USwitch from './packages/u-switch'
-import UCheckbox from './packages/u-checkbox/u-checkbox'
-import UCheckboxs from './packages/u-checkbox/u-checkboxs'
-import URadio from './packages/u-radio/u-radio'
-import URadios from './packages/u-radio/u-radios'
-import UForm from './packages/u-form/u-form'
-import UFormItem from './packages/u-form/u-form-item'
-
-// usage
-import UTableColumn from './packages/u-table/u-table-column'
-import UTable from './packages/u-table/u-table'
-import UPagination from './packages/u-pagination'
-import UTab from './packages/u-tab/u-tab'
-import UTabs from './packages/u-tab/u-tabs'
-import UCrumb from './packages/u-crumb/u-crumb.vue'
-import UCrumbItem from './packages/u-crumb/u-crumb-item.vue'
-
-// popup
-import UModal from './packages/u-modal'
-import UToast from './packages/u-toast'
-import UPopper from './packages/u-popper'
-import UTooltip from './packages/u-tooltip'
-
-import './assets/css/reset.scss'
-
-export {
-    ULink,
-    UButton,
-    UInput,
-    USelect,
-    UCheckbox,
-    UCheckboxs,
-    USwitch,
-    ULayout,
-    UTableColumn,
-    UTable,
-    UPagination,
-    UTab,
-    UTabs,
-    UModal,
-    UToast,
-    UPopper,
-    UTooltip,
-    URadio,
-    URadios,
-    UIcon,
-    UForm,
-    UFormItem,
-    UCrumb,
-    UCrumbItem
-}
