+++ src/main.js
@@ -3,8 +3,11 @@ import app from './app.vue'
 import router from './router'
 import 'highlight.js/styles/atom-one-dark.css'
 
+// import YIUI from '../lib'
+// Vue.use(YIUI)
+
 // 注册全局组件
-import * as YIUI from './entry'
+import * as YIUI from '../packages'
 Object.keys(YIUI).forEach(key => {
     Vue.component(key, YIUI[key])
     if (key === 'UToast') {
