+++ src/pages/demo.vue
@@ -208,7 +208,7 @@
 </template>
 
 <script>
-import DComponentItem from '@/components/demo/d-component-item'
+import DComponentItem from '@/components/d-component-item'
 
 export default {
     components: { DComponentItem },
