+++ vue.config.js
@@ -1,14 +1,22 @@
+const path = require('path')
+function resolve(dir) {
+    return path.join(__dirname, dir)
+}
+
 module.exports = {
     publicPath: process.env.site ? '/yi-ui' : '/',
     outputDir: process.env.site ? 'site' : 'dist',
     css: {
         loaderOptions: {
             sass: {
-                data: `@import "@/assets/css/vars.scss";`
+                data: `@import "@packages/assets/css/vars.scss";`
             }
         }
     },
     chainWebpack: config => {
+        // set alias
+        config.resolve.alias.set('@packages', resolve('packages'))
+
         // markdown Loader
         config.module
             .rule('md')
